python-service/requirements.txt
flask>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pytesseract>=0.3.0
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.0.0
numpy>=1.24.0
